pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "dvc>=3.66.1",
    "fastapi>=0.135.1",
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "mlflow>=1.27.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "prophet>=1.3.0",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "snappy>=3.3.2",
    "typer>=0.24.1",
    "uvicorn>=0.41.0",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]
